modularize routes declaration for webant.api
thanks to the utilities introduced in `webant/util.py`
it is now passible to split routes declaration into multiple module

This commit moved archivant related api routes in its own module.
